// servidor-local/src/utils/types.d.ts
export declare enum Role {
    CLIENTE = "cliente",
    ADMIN = "admin",
    PRESTADOR = "prestador",
    ADMIN = "administrador",
    EMPRESA = "empresa"
}
// ...
export interface userType {
    id: string;
    nome: string;
    numero_identidade: string;
    data_nascimento: string;
    email: string;
    password: string;
    telefone: string;
    pais: string;
    localidade: string;
    role: Role;
    enebled: boolean;
    created_at: Date;
    update_at: Date;
}
// ...
export interface UserDBType {
    id: string;
    nome: string;
    numero_identidade: string;
    data_nascimento: string;
    email: string;
    password: string;
    telefone: string;
    pais: string;
    localidade: string;
    role: Role;
    enebled: boolean;
    created_at: Date;
    update_at: Date;
}